Preserve full-width allocator addresses
Widen the object-pool alignment mask before complementing it and keep heap endpoints in uintptr_t. Print full addresses with %p outside matching builds while preserving the retail diagnostic in MUST_MATCH.

Verified the exact main.dol hash and 100% matching report, matching/nonmatching MWCC builds, Clang syntax and tidy, and native 64-bit tests of the actual allocator sources with stubbed platform services.

+++ src/melee/lb/lbheap.c
@@ -80,7 +80,7 @@ void lbHeap_80015900(void)
     s32 create_i;
     struct Heap* main_heap;
     s32 bounds_i;
-    s32 end;
+    uintptr_t end;
     struct Heap* bounds_heap;
     uintptr_t aram_lo;
     uintptr_t aram_hi;
